Fix Linux warning and error when compiling.

+++ iGameCore/IO/VTK XML/iGameVTUReader.cpp
@@ -262,7 +262,7 @@ bool iGameVTUReader::ReadPointAttribute() {
                     else /* Int64*/ {
                         LongLongArray::Pointer arr = LongLongArray::New();
                         arr->SetDimension(scalarComponents);
-                        ReadBase64EncodedArray<long long>(m_Header_8_byte_flag, data_p, arr);
+                        ReadBase64EncodedArray<int64_t>(m_Header_8_byte_flag, data_p, arr);
                         array = arr;
                     }
                 }
@@ -445,7 +445,7 @@ bool iGameVTUReader::ReadCellData() {
                     else /* Int64*/ {
                         LongLongArray::Pointer arr = LongLongArray::New();
                         arr->SetDimension(scalarComponents);
-                        ReadBase64EncodedArray<long long>(m_Header_8_byte_flag, data_p, arr);
+                        ReadBase64EncodedArray<int64_t >(m_Header_8_byte_flag, data_p, arr);
                         array = arr;
                     }
                 }
@@ -556,7 +556,7 @@ ArrayObject::Pointer iGameVTUReader::ReadCellConnectivity() {
             }
             else /* Int64*/ {
                 LongLongArray::Pointer arr = LongLongArray::New();
-                ReadBase64EncodedArray<long long>(m_Header_8_byte_flag, data_p, arr);
+                ReadBase64EncodedArray<int64_t>(m_Header_8_byte_flag, data_p, arr);
                 CellConnects = arr;
             }
         }
@@ -633,7 +633,7 @@ ArrayObject::Pointer iGameVTUReader::ReadCellOffsets() {
             else /* Int64*/ {
                 LongLongArray::Pointer arr = LongLongArray::New();
                 arr->AddValue(0);
-                ReadBase64EncodedArray<long long>(m_Header_8_byte_flag, data_p, arr);
+                ReadBase64EncodedArray<int64_t>(m_Header_8_byte_flag, data_p, arr);
                 CellOffsets = arr;
             }
         }
